pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognition-engine"
version = "0.4.0"
description = "Google ADK-aligned cognition engine and runtime skeleton."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.33.1",
    "pydantic>=2.13.0",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.24.2",
    "pandas>=3.0.2",
    "numpy>=2.4.4",
    "networkx>=3.6.1",
    "matplotlib>=3.10.8",
    "pygithub>=2.9.1",
    "discord.py>=2.7.1",
    "gitpython>=3.1.47",
    "google-adk>=2.0.0b1,<2.1",
    "litellm==1.82.6",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
]

[project.optional-dependencies]
release = [
    "build>=1.4.4",
    "twine>=6.2.0",
]

test = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.8.0",
    "pytest-timeout>=2.4.0",
    "factory-boy>=3.3.3",
    "Faker>=40.15.0",
    "freezegun>=1.5.5",
    "hypothesis>=6.152.1",
    "pytest-html>=4.2.0",
    "pytest-json-report>=1.5.0",
]

[project.scripts]
ce = "cognition_engine.cli:main"

[tool.setuptools.packages.find]
include = [
    "cognition_engine*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = [".git", ".venv", "venv", "env", "build", "dist", "*.egg-info"]
markers = [
    "unit: 单元测试",
    "integration: 集成测试",
    "slow: 慢速测试（运行时间 > 5秒）",
    "network: 需要网络连接的测试",
    "mock: 使用模拟的测试",
]
addopts = ["-v", "--tb=short", "--strict-markers"]
